requirements.txt
app-version==1.0.1
Babel==2.6.0
dj-database-url==0.5.0
Django==2.2
django-appconf==1.0.3
django-cors-headers==3.0.2
django-extensions==2.1.6
django-heroku==0.3.1
django-permission==1.0.4
djangorestframework==3.9.2
gunicorn==19.9.0
Pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
python-decouple==3.1
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.4
